Move interceptor middleware to a separate class
